Improve channel ban logic to ensure sender chat ID is not the same as the chat ID

+++ handlers/admin/all-chanell-ban.js
@@ -17,7 +17,10 @@ composer.hears('!banChannel', onlyAdmin, async (ctx, next) => {
 })
 
 composer.on('message', async (ctx, next) => {
-  if (ctx.message.sender_chat && ctx.group.info.settings.banChannel && !ctx.message.is_automatic_forward) {
+  if (ctx.message.sender_chat &&
+      ctx.group.info.settings.banChannel &&
+      !ctx.message.is_automatic_forward &&
+      ctx.message.sender_chat.id !== ctx.chat.id) {
     await ctx.deleteMessage()
 
     return ctx.tg.callApi('banChatSenderChat', {
